+++ client/app/templates/_template-editor.sass
@@ -1,9 +1,7 @@
 @import 'app_colors'
 
-.template-content
+.ace-editor
 
-  border: 1px solid $color-gray-border
-
-  .ace-editor
     height: 300px
     width: 100%
+    border: 1px solid $color-gray-border
