+++ client/app/templates/template-editor.html
@@ -45,7 +45,7 @@
                 <div>
                   <label class="control-label col-sm-4">{{'Draft' | translate}}</label>
                   <div class="col-sm-8">
-                      <input type="checkbox" ng-model="vm.template.draft" ng-true-value="true" ng-false-value="false" />
+                      <input type="checkbox" ng-model="vm.template.draft" />
                   </div>
                 </div>
               </div>
@@ -56,7 +56,7 @@
               <h3 translate>Code</h2>
             </div>
           </div>
-          <div ace-editor content="vm.template.content" class="template-content"></div>
+          <ace-editor content="vm.template.content"></ace-editor>
         </div>
         </form>
       </section>
@@ -66,7 +66,7 @@
     <div class="panel panel-default ss-details-panel ss-details-panel-bottom">
       <div class="panel-body">
         <section class="ss-details-section pull-left">
-          <button class="btn btn-primary" type="button" ng-click="vm.saveTemplate()" ng-disabled="!vm.isPristine()" translate>Save</button>
+          <button class="btn btn-primary" type="button" ng-click="vm.saveTemplate()" ng-disabled="templateForm.$pristine" translate>Save</button>
           <button class="btn btn-default" type="button" translate ng-click="vm.cancelChanges()">Cancel</button>
         </section>
       </div>
